Fixed RandomPicker::take
Each time this function was called, it would increase the probability of
the entry added after the chosen entry. This function was only used by
the WangFiller, so it might have affected the randomness with which
different options were used.

Issue resolved by storing the added entries in a vector in an
alternative implementation now called RandomTaker.