Added property names to undo commands
* Set Property 'foo'
* Add Property 'foo'
* Remove Property 'foo'

Currently I haven't touched "Rename Property" because I think it would
get a little too verbose if we mentioned two names.

However, the name does mention the full path of a property if it is
nested, separated by dots.

Also addressed a old occurrence of 'Type' which should have been renamed
to 'Class' in Tiled 1.9.